requirements.txt
aiofiles==0.7.0
anyio==3.3.4
asgiref==3.4.1
attrs==21.2.0
awsebcli==3.20.2
backports.entry-points-selectable==1.1.1
bcrypt==3.2.0
blessed==1.19.0
cached-property==1.5.2
cement==2.8.2
certifi==2021.10.8
cffi==1.15.0
charset-normalizer==2.0.7
click==8.0.3
colorama==0.4.3
cryptography==35.0.0
defusedxml==0.7.1
Deprecated==1.2.13
diff-match-patch==20200713
distlib==0.3.3
dj-database-url==0.5.0
Django==3.2.4
django-allauth==0.45.0
django-cors-headers==3.10.0
django-crispy-forms==1.11.2
django-heroku==0.3.1
django-import-export==2.5.0
django-oauth-toolkit==1.5.0
django-phonenumber-field==5.2.0
django-rest-auth==0.9.5
django-storages==1.12.2
djangorestframework==3.12.4
dockerpty==0.4.1
docopt==0.6.2
drf-social-oauth2==1.0.9
ecdsa==0.17.0
et-xmlfile==1.1.0
filelock==3.3.2
future==0.16.0
greenlet==1.1.2
gunicorn==20.1.0
h11==0.12.0
idna==3.3
importlib-metadata==4.8.1
importlib-resources==5.3.0
iniconfig==1.1.1
jmespath==0.10.0
jsonschema==3.2.0
jwcrypto==1.0
Mako==1.1.5
MarkupPy==1.14
MarkupSafe==2.0.1
mock==4.0.3
multidict==4.7.6
oauthlib==3.1.1
odfpy==1.4.1
openpyxl==3.0.7
packaging==21.0
paramiko==2.8.0
pathspec==0.5.9
phonenumbers==8.12.24
Pillow==8.4.0
platformdirs==2.4.0
pluggy==1.0.0
psycopg2-binary==2.8.6
py==1.10.0
pyasn1==0.4.8
pycparser==2.20
pydantic==1.8.2
PyJWT==2.3.0
PyNaCl==1.4.0
pyparsing==2.4.7
pyrsistent==0.18.0
pysqlite3==0.4.6
pytest==6.2.5
python-dateutil==2.8.2
python-decouple==3.5
python-dotenv==0.19.1
python-http-client==3.3.3
python-jose==3.3.0
python-multipart==0.0.5
python-social-auth==0.3.6
python3-openid==3.2.0
pytz==2021.1
PyYAML==5.4.1
qrcode==7.3.1
requests==2.26.0
requests-oauthlib==1.3.0
rsa==4.7.2
s3transfer==0.5.0
semantic-version==2.8.5
six==1.14.0
sniffio==1.2.0
social-auth-app-django==5.0.0
social-auth-core==4.1.0
starkbank-ecdsa==1.1.1
starlette==0.16.0
stripe==2.61.0
tablib==3.0.0
termcolor==1.1.0
texttable==1.6.4
toml==0.10.2
typing-extensions==3.10.0.2
urllib3==1.26.7
uvicorn==0.15.0
virtualenv==20.10.0
wcwidth==0.1.9
websocket-client==0.59.0
whitenoise==5.3.0
wrapt==1.13.3
xlrd==2.0.1
xlwt==1.3.0
zipp==3.6.0
